Revert gallop code
There were some very rare bugs in the logic that would cause the
tim_sort_merge_left function to go out of bounds. It is also possible
that it was causing the sort to be non-stable.

Thanks to Emanuel Falkenauer for finding the bug.

I'm reverting the gallop code for now due to the extra complexity. We
can add it back in later if we can possibly simplify and verify the
logic carefully.

cc @liwt31